# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Chainsaw Manual Training App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /modules:
    get:
      operationId: listModules
      tags: [modules]
      summary: List all training modules with lock status for the current user
      parameters:
        - name: deviceId
          in: header
          required: true
          schema:
            type: string
        - name: activationCode
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
  /modules/{moduleId}:
    get:
      operationId: getModule
      tags: [modules]
      summary: Get a specific module with video details
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
        - name: deviceId
          in: header
          required: true
          schema:
            type: string
        - name: activationCode
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Module detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleDetail"
        "403":
          description: Module locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Module:
      type: object
      required: [id, title, description, order, isLocked, isCompleted, duration, category, contentType]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        order:
          type: integer
        category:
          type: string
        subCategory:
          type: ["string", "null"]
        contentType:
          type: string
        isLocked:
          type: boolean
        isCompleted:
          type: boolean
        quizPassed:
          type: boolean
        duration:
          type: integer
        thumbnailUrl:
          type: ["string", "null"]
        isHighRisk:
          type: boolean
        pdfUrl:
          type: ["string", "null"]
        learningOutcome:
          type: ["string", "null"]
        assessmentCriteria:
          type: ["string", "null"]
    ModuleDetail:
      type: object
      required: [id, title, description, order, isLocked, isCompleted, duration, vimeoId, category, contentType]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        order:
          type: integer
        category:
          type: string
        subCategory:
          type: ["string", "null"]
        contentType:
          type: string
        isLocked:
          type: boolean
        isCompleted:
          type: boolean
        quizPassed:
          type: boolean
        duration:
          type: integer
        vimeoId:
          type: string
        pdfUrl:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        isHighRisk:
          type: boolean
        lastTimestamp:
          type: ["integer", "null"]
        safetyText:
          type: ["string", "null"]
        learningOutcome:
          type: ["string", "null"]
        assessmentCriteria:
          type: ["string", "null"]
        quizCount:
          type: integer
...